Reorganize package structure for protocol-based architecture
Move core abstractions into probpipe/core/:
- distribution.py (from distributions/) – base hierarchy lives with node.py
- provenance.py (from top-level) – DAG utilities co-located with core

Flatten internal utility modules:
- array_backend/utils.py → _array_utils.py (delete array_backend/ subpackage)
- utils.py → _utils.py (underscore prefix signals internal use)

Remove surrogate/ subpackage:
- EmulatorMixin incompatible with upcoming immutable distribution design
- Remove EmulatorMixin tests and API doc reference

Update all internal imports across 35 files. Re-execute affected notebooks
to regenerate clean outputs. Public API (probpipe.__init__) unchanged
except for EmulatorMixin removal.
