--- probpipe/array_backend/__init__.py
@@ -1,1 +1,0 @@
-"""Array canonicalization utilities for probpipe. Internal use only."""
