+++ probpipe/core/modeling.py
@@ -18,7 +18,7 @@ import jax.numpy as jnp
 import tensorflow_probability.substrates.jax.mcmc as tfp_mcmc
 
 from ..custom_types import Array, ArrayLike, PRNGKey
-from ..distributions.distribution import ArrayDistribution, EmpiricalDistribution, Provenance
+from .distribution import ArrayDistribution, EmpiricalDistribution, Provenance
 from ..distributions.multivariate import MultivariateNormal
 from .node import AbstractModule, Module, WorkflowFunction, abstractwf, wf
