+++ probpipe/linalg/linear_operator.py
@@ -15,7 +15,7 @@ from abc import ABC, abstractmethod
 import math
 
 from ..custom_types import Array, ArrayLike
-from ..array_backend.utils import (
+from .._array_utils import (
     _ensure_real_scalar,
     _ensure_vector,
     _ensure_matrix,
