--- probpipe/surrogate/__init__.py
@@ -1,7 +1,0 @@
-"""Surrogate modeling utilities for ProbPipe."""
-
-from probpipe.surrogate.emulator_mixin import EmulatorMixin
-
-__all__ = [
-    "EmulatorMixin",
-]
