+++ tests/test_sequential_joint.py
@@ -14,7 +14,7 @@ from probpipe import (
     EmpiricalDistribution,
     JointDistribution,
 )
-from probpipe.distributions.distribution import PyTreeArrayDistribution
+from probpipe.core.distribution import PyTreeArrayDistribution
 from probpipe.core.node import WorkflowFunction
